third_party/WebKit/LayoutTests/editing/input/ctrl-up-down-expected.txt

PASS selection.type is "Caret"
PASS selection.focusNode is paragraph1
PASS selection.focusOffset is 0
PASS selection.type is "Caret"
PASS selection.focusNode is paragraph2
PASS selection.focusOffset is 3
PASS selection.type is "Range"
PASS selection.focusNode is paragraph2
PASS selection.focusOffset is 84
PASS selection.type is "Range"
PASS selection.focusNode is paragraph1
PASS selection.focusOffset is 3
PASS successfullyParsed is true

TEST COMPLETE
